Returned app to its default template so that code splitting works again.
